+++ projects/utils-maps/src/public-api.ts
@@ -3,6 +3,9 @@
  */
 
 export * from './lib/ol/utils-ol';
+export * from './lib/ol/customRenderers/dtm_renderer';
+export * from './lib/ol/customRenderers/interpolation_renderer';
+export * from './lib/ol/customRenderers/particle_renderer';
 export * from './lib/webgl/webgl';
 export * from './lib/webgl/engine.core';
 export * from './lib/webgl/engine.helpers';